+++ RELEASE_NOTES.md
@@ -3,6 +3,7 @@
 ### TBD ###
 
 * Adds `rating` element
+* Adds `endorsement` element
 
 ### 1.3.3 2017/05/16 ##